+++ src/main/java/com/ddiring/backend_asset/entitiy/Bank.java
@@ -27,7 +27,7 @@ public class Bank {
     private String role;
 
     @Column(name = "deposit", nullable = false)
-    private Integer deposit;
+    private Long deposit;
 
     @Column(name = "linked_at")
     private LocalDate linkedAt;
@@ -45,7 +45,7 @@ public class Bank {
     private LocalDate updatedAt; // Integer -> LocalDate로 수정
 
     @Builder
-    public Bank(String userSeq, Integer deposit, String bankNumber, String role, LocalDate linkedAt, Integer createdId, LocalDate createdAt, Integer updatedId, LocalDate updatedAt) {
+    public Bank(String userSeq, Long deposit, String bankNumber, String role, LocalDate linkedAt, Integer createdId, LocalDate createdAt, Integer updatedId, LocalDate updatedAt) {
         this.userSeq = userSeq;
         this.bankNumber = bankNumber;
         this.role = role;
