+++ src/main/java/com/ddiring/backend_asset/entitiy/EscrowHistory.java
@@ -33,7 +33,7 @@ public class EscrowHistory {
     private String title;
 
     @Column(name = "price", nullable = false)
-    private Integer price;
+    private Long price;
 
     @Column(name = "transfer_type", nullable = false)
     private Integer transferType;
@@ -42,7 +42,7 @@ public class EscrowHistory {
     private LocalDate transferDate;
 
     @Builder
-    public EscrowHistory(String userSeq, String role, String escrowAccount, String title, Integer price, Integer transferType, LocalDate transferDate) {
+    public EscrowHistory(String userSeq, String role, String escrowAccount, String title, Long price, Integer transferType, LocalDate transferDate) {
         this.userSeq = userSeq;
         this.escrowAccount = escrowAccount;
         this.role = role;
